Move add esim plan out of OnBoardingCoordinator to ESimManager.
It's not the responsibility of the OnboardingCoordinator to know the internals of how to add a plan.
